doModify, doDelete에 로그인 체크, 권한 체크

+++ src/main/java/com/kkwo/demo/controller/UsrArticleController.java
@@ -65,23 +65,63 @@ public class UsrArticleController {
 	
 	@RequestMapping("/usr/article/modify")
 	@ResponseBody
-	public ResultData<Integer> doModify(int id, String title, String body){
+	public ResultData<Integer> doModify( HttpSession httpSession, int id, String title, String body){
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(!isLogined) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+		
 		Article article = articleService.getArticle(id); 
+		
 		if(article == null) {
 			return ResultData.from("F-3", Ut.f("%d글은 존재하지 않습니다", id), id);
 		}
+		
+		if(article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-3", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
+		}
+		
 		articleService.modifyArticle(id, title, body);
+		
 		return ResultData.from("S-1", Ut.f("%d번 글을 수정했습니다", id), id);
 	}
 	
 	@RequestMapping("/usr/article/delete")
 	@ResponseBody
-	public ResultData<Integer> doDelete(int id){
+	public ResultData<Integer> doDelete( HttpSession httpSession, int id){
+		
+		boolean isLogined = false;
+		int loginedMemberId = 0;
+		
+		if(httpSession.getAttribute("loginedMemberId") != null) {
+			isLogined = true;
+			loginedMemberId = (int) httpSession.getAttribute("loginedMemberId");
+		}
+		
+		if(!isLogined) {
+			return ResultData.from("F-A", "로그인 후 이용해주세요");
+		}
+		
 		Article article = articleService.getArticle(id); 
+		
 		if(article == null) {
 			return ResultData.from("F-4", Ut.f("%d번 글은 존재하지 않습니다", id), id);
 		}
+		
+		if(article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-4", Ut.f("%d번 글에 대한 권한이 없습니다", id), id);
+		}
+		
 		articleService.deleteArticle(id);
+		
 		return ResultData.from("S-1", Ut.f("%d번 글을 삭제했습니다", id), id);
 	}
 }
